Tidy dataset definition and add comments
Moving all functions into `variable_lib.py`.
Restructure code in `dataset_definition.py` using this structure:

- Preprocessing data for later use
- Define and extract dataset variables
- Define dataset restrictions
- Apply dataset restrictions and define study population
