+++ analysis/variable_lib.py
@@ -4,6 +4,22 @@ from functools import reduce
 from databuilder.codes import ICD10Code
 from databuilder.ehrql import case, when
 from databuilder.tables.beta import tpp as schema
+from databuilder.codes import Codelist
+
+
+def has_prior_event(prior_events, codelist, where=True):
+    return (
+        prior_events.take(where)
+        .take(prior_events.snomedct_code.is_in(codelist))
+        .exists_for_patient()
+    )
+
+
+def combine_codelists(*codelists):
+    codes = set()
+    for codelist in codelists:
+        codes.update(codelist.codes)
+    return Codelist(codes=codes, category_maps={})
 
 
 def any_of(conditions):
